requirements.txt
flask
picamera2
PyYAML
Werkzeug
Flask-Limiter